attempt to fix linking (happening when constructing gsMaterialMatrixLinear inside a cpp file): - Make includes in gsMaterialMatrix-classes more uniform - Move XML reads/writes for material matrices (gives problems!! where to put them)
Breaking change:
- Changed getMaterialMatrix to uPtr to prevent user-deletion
