+++ examples/example_shell3D_DWR.cpp
@@ -281,7 +281,7 @@ int main(int argc, char *argv[])
     std::vector<gsFunctionSet<>*> parameters(2);
     parameters[0] = &E;
     parameters[1] = &nu;
-    gsMaterialMatrixBase<real_t>* materialMatrix;
+    gsMaterialMatrixBase<real_t>::uPtr materialMatrix;
     gsOptionList options;
     options.addInt("Material","Material model: (0): SvK | (1): NH | (2): NH_ext | (3): MR | (4): Ogden",0);
     options.addInt("Implementation","Implementation: (0): Composites | (1): Analytical | (2): Generalized | (3): Spectral",1);
@@ -660,7 +660,6 @@ int main(int argc, char *argv[])
     }
 
 
-    delete materialMatrix;
     return EXIT_SUCCESS;
 
 }// end main
