+++ examples/example_shell3D_DWR_buckling.cpp
@@ -265,7 +265,7 @@ int main(int argc, char *argv[])
             }
         }
         std::sort(gammas_an.begin(),gammas_an.end());
-        
+
         gammas_num = std::vector<real_t>{0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
 
         lambda_an = gammas_an[modeIdx] / (Load);
@@ -283,7 +283,7 @@ int main(int argc, char *argv[])
     std::vector<gsFunctionSet<> *> parameters(2);
     parameters[0] = &E;
     parameters[1] = &nu;
-    gsMaterialMatrixBase<real_t> *materialMatrix;
+    gsMaterialMatrixBase<real_t>::uPtr materialMatrix;
     gsOptionList options;
     options.addInt("Material", "Material model: (0): SvK | (1): NH | (2): NH_ext | (3): MR | (4): Ogden", 0);
     options.addInt("Implementation", "Implementation: (0): Composites | (1): Analytical | (2): Generalized | (3): Spectral", 1);
@@ -470,7 +470,7 @@ int main(int argc, char *argv[])
         else if (solverH.info()==Spectra::CompInfo::NotConverging) { GISMO_ERROR("Spectra did not converge! Error code: NotConverging"); }
         else if (solverH.info()==Spectra::CompInfo::NotComputed)   { GISMO_ERROR("Spectra did not converge! Error code: NotComputed");   }
         else                                                      { GISMO_ERROR("No error code known"); }
-#else        
+#else
         gsEigen::GeneralizedSelfAdjointEigenSolver< typename gsMatrix<>::Base >  solverH;
         solverH.compute(Kdiff,K_L);
 #endif
@@ -613,7 +613,6 @@ int main(int argc, char *argv[])
         gsWriteParaview<>( fieldPL, "primalL", 1000);
     }
 
-    delete materialMatrix;
     return EXIT_SUCCESS;
 
 } // end main
