+++ examples/gsMaterialMatrix_test.cpp
@@ -67,7 +67,7 @@ int main (int argc, char** argv)
     else
         GISMO_ERROR("Material unknown");
 
-    gsMaterialMatrixBase<real_t>* materialMatrix;
+    gsMaterialMatrixBase<real_t>::uPtr materialMatrix;
     gsOptionList options;
     options.addInt("Material","Material model: (0): SvK | (1): NH | (2): NH_ext | (3): MR | (4): Ogden",material);
     options.addSwitch("Compressibility","Compressibility: (false): Imcompressible | (true): Compressible",Compressibility);
