+++ gsKLShell.h
@@ -1,7 +1,5 @@
 #include <gsCore/gsTemplateTools.h>
 
-#include <gsKLShell/src/getMaterialMatrix.h>
-
 #include <gsKLShell/src/gsMaterialMatrixBase.h>
 #include <gsKLShell/src/gsMaterialMatrixBaseDim.h>
 #include <gsKLShell/src/gsMaterialMatrixComposite.h>
@@ -10,6 +8,7 @@
 #include <gsKLShell/src/gsMaterialMatrixLinear.h>
 #include <gsKLShell/src/gsMaterialMatrixNonlinear.h>
 #include <gsKLShell/src/gsMaterialMatrixTFT.h>
+#include <gsKLShell/src/gsMaterialMatrixContainer.h>
 #include <gsKLShell/src/gsMaterialMatrixUtils.h>
 #include <gsKLShell/src/gsMaterialMatrixXml.h>
 
@@ -18,6 +17,8 @@
 
 #include <gsKLShell/src/gsThinShellFunctions.h>
 #include <gsKLShell/src/gsThinShellUtils.h>
+
+#include <gsKLShell/src/getMaterialMatrix.h>
 
 namespace gismo
 {
