+++ src/gsMaterialMatrixBase_.cpp
@@ -2,15 +2,10 @@
 
 #include <gsKLShell/src/gsMaterialMatrixBase.h>
 #include <gsKLShell/src/gsMaterialMatrixBase.hpp>
-#include <gsKLShell/src/gsMaterialMatrixBaseDim.h>
-#include <gsKLShell/src/gsMaterialMatrixBaseDim.hpp>
 
 namespace gismo
 {
   CLASS_TEMPLATE_INST gsMaterialMatrixBase<real_t>;
-
-  CLASS_TEMPLATE_INST gsMaterialMatrixBaseDim<2,real_t>;
-  CLASS_TEMPLATE_INST gsMaterialMatrixBaseDim<3,real_t>;
 
   #ifdef GISMO_WITH_PYBIND11
