+++ src/gsMaterialMatrixContainer.h
@@ -46,142 +46,48 @@ public:
 public:
 
     /// Constructor
-    gsMaterialMatrixContainer( index_t size = 0 )
-    {
-        m_container.resize(size);
-        // To do: initialize with null pointers
-    }
+    gsMaterialMatrixContainer( index_t size = 0 );
 
-    // gsMaterialMatrixContainer(const gsMaterialMatrixContainer & other)
-    // {
-    //     // for (index_t k=0; k!=other.m_container.size(); k++)
-    //     //     add(memory::make_unique(other.m_container.at(k)));
-    //     m_container = give(other.m_container);
-    // }
-
-    ~gsMaterialMatrixContainer()
-    {
-        // freeAll(m_container);
-    }
+    ~gsMaterialMatrixContainer();
 
     /// Add a material matrix by copying argument
-    void add(const gsMaterialMatrixBase<T> & mat)
-    {
-        m_container.push_back( memory::make_shared(mat.clone().release()) );
-    }
+    void add(const gsMaterialMatrixBase<T> & mat);
 
-    ///\brief Add a material matrix from a gsMaterialMatrixBase<T>::uPtr
-    void add(const gsMaterialMatrixBase<T> * mat)
-    {
-        m_container.push_back( memory::make_shared_not_owned(mat) );
-    }
+    ///\brief Add a material matrix from a gsMaterialMatrixBase<T> *
+    void add(const gsMaterialMatrixBase<T> * mat);
 
     /// Set a material matrix by copying argument
-    void set(const index_t i, const gsMaterialMatrixBase<T> & mat)
-    {
-        m_container[i] = memory::make_shared(mat.clone().release());
-    }
+    void set(const index_t i, const gsMaterialMatrixBase<T> & mat);
+
+    ///\brief Set a material matrix from a gsMaterialMatrixBase<T> *
+    void set(const index_t i, const gsMaterialMatrixBase<T> * mat);
 
     ///\brief Set a material matrix from a gsMaterialMatrixBase<T>::uPtr
-    void set(const index_t i, const gsMaterialMatrixBase<T> * mat)
-    {
-        m_container[i] = memory::make_shared_not_owned(mat);
-    }
+    void set(const index_t i, const typename gsMaterialMatrixBase<T>::Ptr mat);
 
-    ///\brief Set a material matrix from a gsMaterialMatrixBase<T>::uPtr
-    void set(const index_t i, const typename gsMaterialMatrixBase<T>::Ptr mat)
-    {
-        m_container[i] = mat;
-    }
-
-    gsMaterialMatrixBase<T> * piece(const index_t k) const
-    {
-        return m_container.at(k).get();
-    }
+    gsMaterialMatrixBase<T> * piece(const index_t k) const;
 
     index_t size() const {return m_container.size();}
 
-    std::ostream &print(std::ostream &os) const
-    {
-        os << "Piecewise Function with "<<m_container.size() <<" pieces.\n";
-        return os;
-    }
+    std::ostream &print(std::ostream &os) const;
 
-    friend std::ostream & operator<<(std::ostream & os, const gsMaterialMatrixContainer & pwf)
+    friend std::ostream & operator<<(std::ostream & os, const gsMaterialMatrixContainer & mc)
     {
-        return pwf.print(os);
+        return mc.print(os);
     }
 
     /// Clear all function pointers
-    void clear()
-    {
-        m_container.clear();
-    }
+    void clear();
 
 protected:
     Container m_container;
 
 };
 
-namespace internal
-{
-
-/// @brief get a MaterialMatrixContainer from XML data
-///
-/// \ingroup KLShell
-template<class T>
-class gsXml< gsMaterialMatrixContainer<T> >
-{
-private:
-    gsXml() { }
-    typedef gsMaterialMatrixContainer<T> Object;
-
-public:
-    GSXML_COMMON_FUNCTIONS(gsMaterialMatrixContainer<T>);
-    static std::string tag ()  { return "MaterialMatrixContainer"; }
-    static std::string type () { return ""; }
-
-    GSXML_GET_POINTER(Object);
-
-    static void get_into(gsXmlNode * node,Object & obj)
-    {
-        const int size = atoi(node->first_attribute("size")->value());
-
-        // Read material inventory
-        int count = countByTag("MaterialMatrix", node);
-        std::vector<typename gsMaterialMatrixBase<T>::Ptr> mat(count);
-        for (gsXmlNode * child = node->first_node("MaterialMatrix"); child; child =
-                child->next_sibling("MaterialMatrix"))
-        {
-            const int i = atoi(child->first_attribute("index")->value());
-            mat[i] = memory::make_shared(gsXml<gsMaterialMatrixBase<T>>::get(child));
-        }
-
-        obj = gsMaterialMatrixContainer<T>(size);
-        for (gsXmlNode * child = node->first_node("group"); child;
-                child = child->next_sibling("group"))
-        {
-            const int mIndex = atoi(child->first_attribute("material")->value());
-            std::istringstream group_str;
-            group_str.str( child->value() );
-
-            for(int patch; ( gsGetInt(group_str,patch)); )
-                obj.set(patch,mat[mIndex]);
-        }
-
-    }
-
-    static gsXmlNode * put (const Object & /* obj */,
-                            gsXmlTree & /* data */)
-    {
-        GISMO_ERROR("Writing gsMaterialMatrixContainer to Xml is not implemented");
-        // gsWarn<<"Writing gsMaterialMatrixContainer to Xml is not implemented\n";
-        // gsXmlNode * result;
-        // return result;
-        // return putMaterialMatrixToXml< Object >( obj,data );
-    }
-};
-
-} // namespace internal
 
 } // namespace gismo
+
+
+#ifndef GISMO_BUILD_LIB
+#include GISMO_HPP_HEADER(gsMaterialMatrixContainer.hpp)
+#endif
