+++ src/gsMaterialMatrixEval.h
@@ -30,66 +30,38 @@ public:
     /// Constructor
     gsMaterialMatrixEval(   const gsMaterialMatrixContainer<T> & materialMatrices,
                             const gsFunctionSet<T> * deformed,
-                            const gsMatrix<T> z)
-    :
-    m_materialMatrices(materialMatrices),
-    m_deformed(deformed),
-    m_z(z)
-    {
-        for (index_t p = 0; p!=deformed->nPieces(); ++p)
-            GISMO_ASSERT(materialMatrices.piece(p)->initialized(),"Material matrix "<<p<<" is incomplete!");
-
-        this->_makePieces();
-    }
+                            const gsMatrix<T> z);
 
     /// Constructor
     gsMaterialMatrixEval(   gsMaterialMatrixBase<T> * materialMatrix,
                             const gsFunctionSet<T> * deformed,
-                            const gsMatrix<T> z)
-    :
-    m_materialMatrices(deformed->nPieces()),
-    m_deformed(deformed),
-    m_z(z)
-    {
-        GISMO_ASSERT(materialMatrix->initialized(),"Material matrix is incomplete!");
-        for (index_t p = 0; p!=deformed->nPieces(); ++p)
-            m_materialMatrices.set(p,materialMatrix);
-        this->_makePieces();
-    }
+                            const gsMatrix<T> z);
+
+    /// Constructor
+    gsMaterialMatrixEval(   typename gsMaterialMatrixBase<T>::uPtr & materialMatrix,
+                            const gsFunctionSet<T> * deformed,
+                            const gsMatrix<T> z);
 
     /// Constructor
     gsMaterialMatrixEval(  const gsMaterialMatrixContainer<T> & materialMatrices,
                             const gsFunctionSet<T> * undeformed,
                             const gsFunctionSet<T> * deformed,
-                            const gsMatrix<T> z)
-    :
-    m_materialMatrices(materialMatrices),
-    m_deformed(deformed),
-    m_z(z)
-    {
-        this->_makePieces(undeformed);
-    }
+                            const gsMatrix<T> z);
 
     /// Constructor
     gsMaterialMatrixEval(  gsMaterialMatrixBase<T> * materialMatrix,
                             const gsFunctionSet<T> * undeformed,
                             const gsFunctionSet<T> * deformed,
-                            const gsMatrix<T> z)
-    :
-    m_materialMatrices(deformed->nPieces()),
-    m_deformed(deformed),
-    m_z(z)
-    {
-        for (index_t p = 0; p!=deformed->nPieces(); ++p)
-            m_materialMatrices.set(p,materialMatrix);
-        this->_makePieces(undeformed);
-    }
+                            const gsMatrix<T> z);
+
+    /// Constructor
+    gsMaterialMatrixEval(   typename gsMaterialMatrixBase<T>::uPtr & materialMatrix,
+                            const gsFunctionSet<T> * undeformed,
+                            const gsFunctionSet<T> * deformed,
+                            const gsMatrix<T> z);
 
     /// Destructor
-    ~gsMaterialMatrixEval()
-    {
-        freeAll(m_pieces);
-    }
+    ~gsMaterialMatrixEval();
 
     /// Domain dimension, always 2 for shells
     short_t domainDim() const {return 2;}
@@ -104,29 +76,16 @@ public:
     short_t targetDim() const { return this->piece(0).targetDim(); }
 
     /// Implementation of piece, see \ref gsFunction
-    const gsFunction<T> & piece(const index_t p) const
-    {
-        return *m_pieces[p];
-    }
+    const gsFunction<T> & piece(const index_t p) const { return *m_pieces[p]; }
 
     /// Implementation of eval_into, see \ref gsFunction
     void eval_into(const gsMatrix<T>& /*u*/, gsMatrix<T>& /*result*/) const
     { GISMO_NO_IMPLEMENTATION; }
 
 protected:
-    void _makePieces()
-    {
-        m_pieces.resize(m_deformed->nPieces());
-        for (size_t p = 0; p!=m_pieces.size(); ++p)
-            m_pieces.at(p) = new gsMaterialMatrixEvalSingle<T,out>(p,m_materialMatrices.piece(p),m_deformed,m_z);
-    }
+    void _makePieces();
 
-    void _makePieces(const gsFunctionSet<T> * undeformed)
-    {
-        m_pieces.resize(m_deformed->nPieces());
-        for (size_t p = 0; p!=m_pieces.size(); ++p)
-            m_pieces.at(p) = new gsMaterialMatrixEvalSingle<T,out>(p,m_materialMatrices.piece(p),undeformed,m_deformed,m_z);
-    }
+    void _makePieces(const gsFunctionSet<T> * undeformed);
 
 protected:
     gsMaterialMatrixContainer<T> m_materialMatrices;
