+++ src/gsMaterialMatrixIntegrate.hpp
@@ -29,6 +29,103 @@
 namespace gismo
 {
 
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(const gsMaterialMatrixContainer<T> & materialMatrices,
+                                                            const gsFunctionSet<T> * deformed)
+:
+m_materialMatrices(materialMatrices),
+m_deformed(deformed)
+{
+    for (index_t p = 0; p!=m_materialMatrices.size(); p++)
+    {
+        GISMO_ASSERT(materialMatrices.piece(p)!=nullptr,"Material matrix "<<p<<" is incomplete!");
+        GISMO_ASSERT(materialMatrices.piece(p)!=NULL,"Material matrix "<<p<<" is incomplete!");
+        GISMO_ASSERT(materialMatrices.piece(p)->initialized(),"Material matrix "<<p<<" is incomplete!");
+    }
+
+    this->_makePieces(deformed);
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(gsMaterialMatrixBase<T> * materialMatrix,
+                                                            const gsFunctionSet<T> * deformed)
+:
+m_materialMatrices(deformed->nPieces()),
+m_deformed(deformed)
+{
+    GISMO_ASSERT(materialMatrix->initialized(),"Material matrix is incomplete!");
+    for (index_t p = 0; p!=deformed->nPieces(); ++p)
+        m_materialMatrices.set(p,materialMatrix);
+    this->_makePieces(deformed);
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(typename gsMaterialMatrixBase<T>::uPtr & materialMatrix,
+                                                            const gsFunctionSet<T> * deformed)
+:
+gsMaterialMatrixIntegrate(materialMatrix.get(),deformed)
+{
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(const gsMaterialMatrixContainer<T> & materialMatrices,
+                                                            const gsFunctionSet<T> * undeformed,
+                                                            const gsFunctionSet<T> * deformed)
+:
+m_materialMatrices(materialMatrices),
+m_deformed(deformed)
+{
+    this->_makePieces(undeformed,deformed);
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(gsMaterialMatrixBase<T> * materialMatrix,
+                                                            const gsFunctionSet<T> * undeformed,
+                                                            const gsFunctionSet<T> * deformed)
+:
+m_materialMatrices(deformed->nPieces()),
+m_deformed(deformed)
+{
+    for (index_t p = 0; p!=deformed->nPieces(); ++p)
+    {
+        m_materialMatrices.set(p,materialMatrix);
+        this->_makePieces(undeformed,deformed);
+    }
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::gsMaterialMatrixIntegrate(typename gsMaterialMatrixBase<T>::uPtr & materialMatrix,
+                                                            const gsFunctionSet<T> * undeformed,
+                                                            const gsFunctionSet<T> * deformed)
+:
+gsMaterialMatrixIntegrate(materialMatrix.get(),undeformed,deformed)
+{
+}
+
+template <class T, enum MaterialOutput out>
+gsMaterialMatrixIntegrate<T,out>::~gsMaterialMatrixIntegrate()
+{
+    freeAll(m_pieces);
+}
+
+template <class T, enum MaterialOutput out>
+void gsMaterialMatrixIntegrate<T,out>::_makePieces(const gsFunctionSet<T> * deformed)
+{
+    m_pieces.resize(deformed->nPieces());
+    for (size_t p = 0; p!=m_pieces.size(); ++p)
+        m_pieces.at(p) = new gsMaterialMatrixIntegrateSingle<T,out>(p,m_materialMatrices.piece(p),deformed);
+}
+
+
+template <class T, enum MaterialOutput out>
+void gsMaterialMatrixIntegrate<T,out>::_makePieces(const gsFunctionSet<T> * undeformed, const gsFunctionSet<T> * deformed)
+{
+    m_pieces.resize(m_deformed->nPieces());
+    for (size_t p = 0; p!=m_pieces.size(); ++p)
+        m_pieces.at(p) = new gsMaterialMatrixIntegrateSingle<T,out>(p,m_materialMatrices.piece(p),undeformed,deformed);
+}
+
+
 // Linear material models
 template <class T, enum MaterialOutput out>
 gsMaterialMatrixIntegrateSingle<T,out>::gsMaterialMatrixIntegrateSingle(index_t patch,
