+++ src/gsMaterialMatrixNonlinear.h
@@ -15,7 +15,6 @@
 
 #pragma once
 
-#include <gsKLShell/src/gsMaterialMatrixBase.h>
 #include <gsKLShell/src/gsMaterialMatrixBaseDim.h>
 #include <gsKLShell/src/gsMaterialMatrixUtils.h>
 #include <gsIO/gsOptionList.h>
